add class HttpServer instead of adding server in main with microhttpd - fix wristrotation in output
