+++ src/main.cpp
@@ -4,115 +4,11 @@
 #include "ArmAL5D.hpp"
 #include "TransmitionQueue.hpp"
 #include "log.h"
-#include "servercontroller.hpp"
-#include "resthttp.hpp"
+#include "httpserver.hpp"
 #include <getopt.h>
-#include <microhttpd.h>
-#include <string>
-#include <map>
 
-#define MAXANSWERSIZE   512
-
-using std::map;
-using std::string;
 
 TransmitionQueue* gPtrTQ;
-ServerController server;
-
-const char* badpage = "<html><head><title></title></head><body><h1>A error occur on server</h1></body></html>";
-
-static int send_bad_response(struct MHD_Connection * connection)
-{
-  static char * bad_response = (char *)badpage; 
-  int bad_response_len = strlen(bad_response);
-  int ret;
-  struct MHD_Response *response;
-
-  response = MHD_create_response_from_buffer(bad_response_len,bad_response, MHD_RESPMEM_PERSISTENT);
-  if(response ==0)
-  {
-    return MHD_NO;
-  }
-  ret = MHD_queue_response(connection,MHD_HTTP_OK,response);
-  MHD_destroy_response(response);
-  return ret;
-}
-
-//function callback of function get connection value
-//read value in methode get
-static int get_url_arg(void *cls, MHD_ValueKind kind,const char *key, const char * value)
-{
-  map<string, string> * url_arg = static_cast<map<string , string> *> (cls);
-  if (url_arg->find(key) == url_arg->end()){
-    if(!value)
-    {
-      (*url_arg)[key] = ""; 
-    }
-    else
-    {
-      (*url_arg)[key] = value;
-    }
-  }
-  return MHD_YES;
-}
-
-//function callback when exist a request
-static int AnswerRequest(void* cls, struct MHD_Connection * connection,
-                          const char * url, const char * methode,
-                          const char * version, const char * upload_data,
-                          size_t * upload_data_size, void ** ptr){
-  static int dummy;
-  struct MHD_Response * response;
-  int ret;
-  map<string , string> url_arg;
-  RestHttp apicontrol;
-  string respdata;
-  char* respbuffer;
-  if(0!= strcmp(methode, "GET")){
-    return MHD_NO;
-  }
-  //ptr is a pointer used to save the state of the function callback for the future call by MHD( this function can be called by many times)
-  if(&dummy != *ptr)
-  {
-    *ptr = &dummy;
-    return MHD_YES;
-  }
-  if(0 != *upload_data_size)
-    return MHD_NO;
-  //get values from url with get methode  
-  if(MHD_get_connection_values(connection,MHD_GET_ARGUMENT_KIND,get_url_arg,&url_arg)<0)
-  {
-    return send_bad_response(connection);
-  }
-  //call api to control arm
-  apicontrol.response_rest_request(url,url_arg,respdata,server);
-  *ptr = NULL;
-  respbuffer = (char*) malloc(respdata.size()+1);
-  if(respbuffer == 0)
-  {
-    return MHD_NO;
-  }
-  strncpy(respbuffer, respdata.c_str(), respdata.size() + 1);
-  //create response for resquest
-  response = MHD_create_response_from_buffer(strlen(respbuffer), (void *)respbuffer, MHD_RESPMEM_PERSISTENT);
-  if(response == 0)
-  {
-    if(respbuffer != NULL)
-      free(respbuffer);
-    return MHD_NO;
-  }
-  //send response
-  MHD_add_response_header(response, "Access-Control-Allow-Origin", "*");
-	MHD_add_response_header(response, "Vary", "Origin");
-  ret = MHD_queue_response(connection, MHD_HTTP_OK, response);
-  //destroy response
-  MHD_destroy_response(response);
-  if(respbuffer == 0)
-  {
-    return MHD_NO;
-  }
-  return ret;
-}
 
 void signalHandler (int signum) {
   printf("\nreceived signal %d \n",signum);
@@ -219,20 +115,10 @@ int main(int argc, char *argv[]) {
     }
 
     else if(!strcmp(mode,"server-web")){
-      TransmitionQueue tQueue;
-
       LOG_I("Launching Server");
       // for web service
-      struct MHD_Daemon * d;
-      d = MHD_start_daemon(MHD_USE_THREAD_PER_CONNECTION,
-                          3000, NULL, NULL, &AnswerRequest,
-                          NULL, MHD_OPTION_END);
-      if(d == NULL)
-      {
-        return 1;
-      }
-      (void) getc(stdin);
-      MHD_stop_daemon(d);
+      HttpServer httpserver;
+      httpserver.http_server_run();
       LOG_I("Program end!");
     }
   }
